requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
openai>=1.0
python-dotenv>=1.0
requests>=2.31
